+++ src/main/java/com/bravo/user/service/PasswordService.java
@@ -1,7 +1,8 @@
-package com.bravo.user.dao.model;
+package com.bravo.user.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
 import javax.crypto.BadPaddingException;
 import javax.crypto.Cipher;
@@ -11,34 +12,23 @@ import java.security.InvalidKeyException;
 import java.security.Key;
 import java.util.Base64;
 
-@Component
-public class PasswordEncryptor implements AttributeConverter<String, String> {
+@Service
+public class PasswordService {
 
     private Key key;
     private Cipher cipher;
 
     @Autowired
-    public PasswordEncryptor(Key key, Cipher cipher) {
+    public PasswordService(Key key, Cipher cipher) {
         this.key = key;
         this.cipher = cipher;
     }
 
-    @Override
-    public String convertToDatabaseColumn(String plainText) {
+    public String encrypt(String plainText) {
         try {
             cipher.init(Cipher.ENCRYPT_MODE, key);
             return Base64.getEncoder().encodeToString(cipher.doFinal(plainText.getBytes()));
         } catch (IllegalBlockSizeException | BadPaddingException | InvalidKeyException e) {
-            throw new IllegalStateException(e);
-        }
-    }
-
-    @Override
-    public String convertToEntityAttribute(String encrypted) {
-        try {
-            cipher.init(Cipher.DECRYPT_MODE, key);
-            return new String(cipher.doFinal(Base64.getDecoder().decode(encrypted)));
-        } catch (InvalidKeyException | BadPaddingException | IllegalBlockSizeException e) {
             throw new IllegalStateException(e);
         }
     }
